More GoDaddy API changes
PUT now requires a relative domain name.
GET changed the order of name and data in the JSON.

Note: Due to the API changes, you may need to add
GODADDY_BASE to getssl.cfg.  This is a space-separated
list of base domain names (zones) in which the challenge
responses are entered.  There doesn't seem to be a way
around this that works in all cases, as the GoDaddy API
requires that the zone name be known.