Make a stricter comparison for matching against zone names
This adds a dot in front of the fqdn and the zone name, so the
comparison will only check against full DNS parts of the fqdn and the
zone name. This will prevent to match a fqdn against an unrelated zone
name which overlaps with the wanted zone in the naming.

Fixes #398