Compare version numbers properly
getssl uses cURL's version to determine what command
options are valid.

The previous shortcuts will fail when curl V8.10 is
released.  (8.9 is greater than 8.10).  V8 is planned
for release, in part to avoid a minor version of 100,
which also would fail.

check_version() will compare full or partial version
strings by component, and is true if $1 is at least $2.